+++ docker-compose.yml
@@ -18,6 +18,7 @@ services:
       dockerfile: Dockerfile
     volumes:
       - ./gatling-runner/src/gatling/resources/config:/app/gatling-runner/src/gatling/resources/config
+      - ./gatling-runner/results:/app/results
     ports:
       - "8091:8091"
     depends_on:
