+++ gatling-adapter/Dockerfile
@@ -9,7 +9,7 @@ RUN gradle assemble
 FROM eclipse-temurin:17-jre-alpine
 
 WORKDIR /app
-# Copy the executables from the build stages
+# Copy the executables from the build stage
 COPY --from=build app gatling-adapter
 
 # Run the jar file
