+++ gatling-adapter/build.gradle
@@ -4,24 +4,19 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.3'
 }
 
-dependencies{
+dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
 
-    implementation 'org.scala-lang:scala-library:2.13.11'
-
-    implementation 'com.opencsv:opencsv:3.7'
+    // json annotations
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.2'
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.2'
-    implementation 'org.json:json:20230618'
-
-    // Run gradle task from source code
-    implementation 'org.gradle:gradle-tooling-api:7.3-20210825160000+0000'
 
     // dqlang
     compileOnly 'org.projectlombok:lombok:1.18.28'
     annotationProcessor 'org.projectlombok:lombok:1.18.28'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.3'
+    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.9.3'
 }
 
 repositories {
