+++ gatling-adapter/src/main/java/poc/service/GatlingAdaptationService.java
@@ -1,10 +1,11 @@
-package poc.adapter;
+package poc.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
+import poc.adapter.GatlingAdapter;
 import poc.config.FileConfig;
 import poc.dqlang.loadtest.*;
 import poc.util.ConfigModelCreator;
@@ -12,7 +13,6 @@ import poc.util.MultiLineFileWriter;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.logging.Logger;
 
 @Component
 @Slf4j
@@ -23,9 +23,7 @@ public class GatlingAdaptationService {
 
     @EventListener(ApplicationReadyEvent.class)
     public void start()  {
-        try {
-            this.adapt();
-        }
+        try { this.adapt(); }
         catch (Exception e) {
             log.error("ADAPTATION FAILED: " + e.getMessage());
             e.printStackTrace();
@@ -35,19 +33,14 @@ public class GatlingAdaptationService {
     private void adapt() throws IOException {
         // TODO In Zukunft muss hier die Config über eine Schnittstelle kommen und nicht im Project erzeugt werden
         LoadTestConfiguration configModel = ConfigModelCreator.create();
-        log.info("LOAD TEST CONFIG WAS CREATED");
+        log.info("LOAD TEST CONFIG MODEL WAS CREATED");
 
         List<String> gatlingConfig = adapter.adapt(configModel);
         log.info("LOAD TEST CONFIG WAS ADAPTED");
 
-        String configPath = FileConfig.getAbsoluteGatlingConfigPath();
+        String configPath = FileConfig.getGatlingConfigPath();
         log.info("GATLING CONFIG WILL BE WRITTEN HERE: " + configPath);
         MultiLineFileWriter.write(gatlingConfig, configPath);
-
-        runGatling();
-    }
-
-    private void runGatling() {
-        // TODO
+        log.info("GATLING CONFIG WAS WRITTEN SUCCESSFULLY");
     }
 }
